# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Spring PetClinic
  description: Spring PetClinic Sample Application.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: '1.0'
servers:
  - url: http://localhost:9966/petclinic/api
...
paths:
...
  /owners/{ownerId}/pets:
    post:
      tags:
        - pet
      operationId: addPetToOwner
      summary: Adds a pet to an owner
      description: Records the details of a new pet.
      parameters:
        - name: ownerId
          in: path
          description: The ID of the pet owner.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      requestBody:
        description: The details of the new pet.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetFields'
        required: true
      responses:
        201:
          description: The pet was successfully added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Pet or Owner not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /owners/{ownerId}/pets/{petId}:
    get:
      tags:
        - pet
      operationId: getOwnersPet
      summary: Get a pet by ID
      description: Returns the pet or a 404 error.
      parameters:
        - name: ownerId
          in: path
          description: The ID of the pet owner.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
        - name: petId
          in: path
          description: The ID of the pet.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      responses:
        200:
          description: Pet details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Owner or pet not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    put:
      tags:
        - pet
      operationId: updateOwnersPet

      summary: Update a pet's details
      description: Updates the pet record with the specified details.
      parameters:
        - name: ownerId
          in: path
          description: The ID of the pet owner.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
        - name: petId
          in: path
          description: The ID of the pet.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      requestBody:
        description: The pet details to use for the update.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetFields'
        required: true
      responses:
        204:
          description: Update successful.
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Pet not found for this owner.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /pets:
    get:
      tags:
        - pet
      operationId: listPets
      summary: Lists pet
      description: Returns an array of pet .
      responses:
        200:
          description: Pet types found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /pets/{petId}:
    get:
      tags:
        - pet
      operationId: getPet
      summary: Get a pet by ID
      description: Returns the pet  or a 404 error.
      parameters:
        - name: petId
          in: path
          description: The ID of the pet.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      responses:
        200:
          description: Pet details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Pet not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    put:
      tags:
        - pet
      operationId: updatePet
      summary: Update a pet by ID
      description: Returns the pet or a 404 error.
      parameters:
        - name: petId
          in: path
          description: The ID of the pet.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      requestBody:
        description: The pet
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
        required: true
      responses:
        200:
          description: Pet details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Pet not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    delete:
      tags:
        - pet
      operationId: deletePet
      summary: Delete a pet by ID
      description: Returns the pet or a 404 error.
      parameters:
        - name: petId
          in: path
          description: The ID of the pet.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      responses:
        200:
          description: Pet details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Pet not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  schemas:
    ProblemDetail:
      title: Problem Detail
      description: The schema for all error responses.
      type: object
      properties:
        type:
          title: Type
          description: Full URL that originated the error response.
          type: string
          format: uri
          example: 'http://localhost:9966/petclinic/api/owner'
          readOnly: true
        title:
          title: Title
          description: The short error title.
          type: string
          example: NoResourceFoundException
          readOnly: true
        status:
          title: Status
          description: HTTP status code
          type: integer
          format: int32
          minimum: 400
          maximum: 600
          exclusiveMaximum: true
          example: 500
          readOnly: true
        detail:
          title: Detail
          description: The long error message.
          type: string
          example: 'No static resource api/owner.'
          readOnly: true
        timestamp:
          title: Timestamp
          description: The time the error occurred.
          type: string
          format: date-time
          example: '2024-11-23T13:59:21.3820407Z'
          readOnly: true
        schemaValidationErrors:
          title: Schema validation errors
          description: Validation errors against the OpenAPI schema.
          type: array
          items:
            $ref: '#/components/schemas/ValidationMessage'
      required:
        - type
        - title
        - status
        - detail
        - timestamp
        - schemaValidationErrors
    ValidationMessage:
      title: Validation message
      description: Messages describing a validation error.
      type: object
      properties:
        message:
          title: Message
          description: The validation message.
          type: string
          example: "[Path '/lastName'] Instance type (null) does not match any allowed primitive type (allowed: ['string'])"
          readOnly: true
      required:
        - message
      additionalProperties: true
...
    PetFields:
      title: Pet fields
      description: Editable fields of a pet.
      type: object
      properties:
        name:
          title: Name
          description: The name of the pet.
          type: string
          maxLength: 30
          example: Leo
        birthDate:
          title: Birth date
          description: The date of birth of the pet.
          type: string
          format: date
          example: '2010-09-07'
        type:
          $ref: '#/components/schemas/PetType'
        weight:
          title: Weight
          description: Pet weight in kilograms (nullable)
          type: number
          format: double
          example: 15.75
          nullable: true
      required:
        - name
        - birthDate
        - type
    Pet:
      title: Pet
      description: A pet.
      allOf:
        - $ref: '#/components/schemas/PetFields'
        - type: object
          properties:
            id:
              title: ID
              description: The ID of the pet.
              type: integer
              format: int32
              minimum: 0
              example: 1
              readOnly: true
            ownerId:
              title: Owner ID
              description: The ID of the pet's owner.
              type: integer
              format: int32
              minimum: 0
              example: 1
              readOnly: true
            visits:
              title: Visits
              description: Vet visit bookings for this pet.
              type: array
              items:
                $ref: '#/components/schemas/Visit'
              readOnly: true
          required:
            - id
            - type
            - visits
...
    VisitFields:
      title: Visit fields
      description: Editable fields of a vet visit.
      type: object
      properties:
        date:
          title: Date
          description: The date of the visit.
          type: string
          format: date
          example: '2013-01-01'
        description:
          title: Description
          description: The description for the visit.
          type: string
          minLength: 1
          maxLength: 255
          example: 'rabies shot'
      required:
        - description
    Visit:
      title: Visit
      description: A booking for a vet visit.
      allOf:
        - $ref: '#/components/schemas/VisitFields'
        - type: object
          properties:
            id:
              title: ID
              description: The ID of the visit.
              type: integer
              format: int32
              minimum: 0
              example: 1
              readOnly: true
            petId:
              title: Pet ID
              description: The ID of the pet.
              type: integer
              format: int32
              minimum: 0
              example: 1
          required:
            - id
            - petId
    PetTypeFields:
      title: PetType fields
      description: Editable fields of a pet type.
      type: object
      properties:
        name:
          title: Name
          description: The name of the pet type.
          type: string
          maxLength: 80
          minLength: 1
          example: cat
      required:
        - name
    PetType:
      title: Pet type
      description: A pet type.
      allOf:
        - $ref: '#/components/schemas/PetTypeFields'
        - type: object
          properties:
            id:
              title: ID
              description: The ID of the pet type.
              type: integer
              format: int32
              minimum: 0
              example: 1
          required:
            - id
...